+++ llvm-hs/test/LLVM/Test/Instructions.hs
@@ -618,7 +618,7 @@ tests = testGroup "Instructions" [
                \  %1 = load i32, i32* @0, align 1\n\
                \  ret i32 %1\n\
                \}\n"
-    s <- withContext $ \context -> withModuleFromAST' context mAST moduleLLVMAssembly
+    s <- withContext $ \context -> withModuleFromAST context mAST moduleLLVMAssembly
     s @?= mStr,
     
   testGroup "terminators" [