+++ llvm-hs/test/LLVM/Test/Support.hs
@@ -13,22 +13,9 @@ import LLVM.Context
 import LLVM.Module
 import LLVM.Diagnostic
 
-class FailInIO f where
-  errorToString :: f -> String
-
-failInIO :: FailInIO f => ExceptT f IO a -> IO a
-failInIO = either (fail . errorToString) return <=< runExceptT
-
-instance FailInIO String where
-  errorToString = id
-
-instance FailInIO (Either String Diagnostic) where
-  errorToString = either id diagnosticDisplay
-
 withModuleFromLLVMAssembly' :: Context -> ByteString -> (Module -> IO a) -> IO a
-withModuleFromLLVMAssembly' c s f  = failInIO $ withModuleFromLLVMAssembly c s f
-withModuleFromAST' c a f = failInIO $ withModuleFromAST c a f
-withModuleFromBitcode' c a f = failInIO $ withModuleFromBitcode c ("<string>", a) f
+withModuleFromLLVMAssembly' c s f  = withModuleFromLLVMAssembly c s f
+withModuleFromBitcode' c a f = withModuleFromBitcode c ("<string>", a) f
 
 assertEqPretty :: (Eq a, Show a) => a -> a -> Assertion
 assertEqPretty actual expected = do
@@ -38,7 +25,7 @@ assertEqPretty actual expected = do
 
 strCheckC mAST mStr mStrCanon = withContext $ \context -> do
   a <- withModuleFromLLVMAssembly' context mStr moduleAST
-  s <- withModuleFromAST' context mAST moduleLLVMAssembly
+  s <- withModuleFromAST context mAST moduleLLVMAssembly
   (a,s) `assertEqPretty` (mAST, mStrCanon)
 
 strCheck mAST mStr = strCheckC mAST mStr mStr