+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   of C-style strings
 * 'preferredAlignment' is now a 'Word32' instead of 'Maybe Word32'. To
   recover the old behavior set it to the same value as abiAlignment.
+* Switch from ExceptT to using exceptions.
+  See `LLVM.Exception` for an overview of the exceptions potentially thrown.
 
 ## 4.0.0 (initial release, changes in comparison to llvm-general)